requirements.txt
mcp
clickhouse-connect
fastapi
sse-starlette
uvicorn
python-dotenv
starlette
